+++ src/index.ts
@@ -1,0 +1,1 @@
+// TODO: run the program, given the input
