+++ src/rover.ts
@@ -1,6 +1,4 @@
 export class MarsRover {
-    private input = '';
-    
     private moves = {
         "N": [0, -1],
         "E": [1, 0],
@@ -17,43 +15,7 @@ export class MarsRover {
 
     private directions = ["N", "E", "S", "W"];
 
-    public parseInput( input: string ) {
-        // grid
-        // rovers[]
-        if (!input) {
-            // throw error for invalid input
-            return;
-        }
-
-        const data = input.trim().split('\n');
-        const gridSize = data[0].split(' ').map(item => item * 1);
-
-        const width = gridSize[0];
-        const height = gridSize[1];
-
-        const gridRows = new Array(width).fill('')
-        const grid = new Array(height).fill(gridRows);
-
-        const rovers = [
-            {
-                start: data[1].split(' '),
-                instructions: data[2].split(''),
-            },
-            {
-                start: data[3].split(' '),
-                instructions: data[4].split(''),
-            }
-        ];
-        // coord
-        // instructions
-
-        return {
-            grid,
-            rovers,
-        };
-    }
-
-    private performAction( action: string ) {
+    private performAction(action: string) {
         // TODO: action must be L, R or M
 
         // if action = L or R -> rotate from current facing by traversing the array +/- 1
@@ -64,17 +26,10 @@ export class MarsRover {
         // then move in the current direction
     }
 
-    constructor( input: string ) {
+    constructor(input: string) {
         // parse the input
 
         return this;
-    }
-
-    /**
-     * doThing
-     */
-    public doThing() {
-        return 'result';
     }
 
     public placeRover() {
@@ -90,13 +45,13 @@ export class MarsRover {
     // the lower-left coordinates are assumed to be 0,0
 
     // Each rover has two lines of input:
-        // The first line gives the rover's position, 
-            /* 
-            The position is made up of two integers and a letter separated by spaces, 
-            corresponding to the x and y coordinates and the rover's orientation.
-            */
+    // The first line gives the rover's position, 
+    /* 
+    The position is made up of two integers and a letter separated by spaces, 
+    corresponding to the x and y coordinates and the rover's orientation.
+    */
 
-        //  the second line is a series of instructions telling the rover how to explore the plateau
+    //  the second line is a series of instructions telling the rover how to explore the plateau
 
     // Each rover will be finished sequentially, 
     // which means that the second rover won`t start to move until
